adds default prediction mechanism, optimizes match monitoring depending on time remaining to match start

+++ tot-server-side/src/main/java/lt/verbus/totalisator/controller/dto/PredictionBasicDTO.java
@@ -15,8 +15,8 @@ public class PredictionBasicDTO {
 
     private Long userId;
 
-    private Byte homeScore;
+    private Integer homeScore;
 
-    private Byte awayScore;
+    private Integer awayScore;
 
 }
