+++ tot-server-side/src/main/java/lt/verbus/totalisator/controller/dto/PredictionDTO.java
@@ -19,9 +19,9 @@ public class PredictionDTO {
 
     private MatchDTO match;
 
-    private Byte homeScore;
+    private Integer homeScore;
 
-    private Byte awayScore;
+    private Integer awayScore;
 
     private Float accuracy;
