+++ tot-server-side/src/main/java/lt/verbus/totalisator/controller/dto/PredictionRegistrationDTO.java
@@ -12,8 +12,8 @@ public class PredictionRegistrationDTO {
 
     private Long userId;
 
-    private Byte homeScore;
+    private Integer homeScore;
 
-    private Byte awayScore;
+    private Integer awayScore;
 
 }
